fix: corregir errores en controlador de lista de empaque
- Pasar `req` a métodos privados mover_item_entre_contenedores y
  restar_mover_contenedor_contenedor para que executeTransactionalTask
  funcione correctamente
- Corregir { _id: idItem } por { _id: idItem._id } al actualizar pallet item
- Unificar llamadas a getPallets para pasar session directamente (no { session })
- Pasar argumento `item` a have_lote_GGN_export en agregar item
- Reemplazar ProcessError por Error estándar en los métodos del controlador
- Redirigir ruta put_proceso_aplicaciones_listaEmpaque_moverItems al nuevo
  ListaEmpaqueController en lugar del antiguo ProcesoRepository
