+++ server/controllers/validations.js
@@ -1,23 +1,21 @@
 
-function have_lote_GGN_export(lote, contenedor, item = "") {
+function have_lote_GGN_export(lote, contenedor) {
     try {
-
         // Desestructurar para acceso más limpio y validación rápida
         if (!lote.GGN) return false;
         const { GGN } = lote.predio;
-        const { PAIS_DESTINO } = contenedor.infoContenedor.clienteInfo || {};
+        const PAIS_DESTINO = contenedor.pais_destino || "";
 
         // Validar que exista GGN y tenga datos
         if (!GGN.paises.length) {
             return false;
         }
-
         if (!contenedor.GGN) {
             return true;
         }
 
         return PAIS_DESTINO &&
-            GGN.paises.some(pais => PAIS_DESTINO.includes(pais))
+            GGN.paises.some(pais => pais.equals(PAIS_DESTINO))
 
     } catch (err) {
         console.error("Error en validación GGN:", { error: err.message, predio: lote.predio?.GGN, tipoFruta: item?.tipoFruta });
